Move `createHashedPath` util to blob storage module It's always used in combination with the blob storage, so it should be there. Also it's used by both the DAM and file uploads module, so having it in the DAM module seems wrong.
