+++ packages/api/cms-api/src/blob-storage/backends/blob-storage-backend.service.ts
@@ -1,9 +1,9 @@
 import { Inject, Injectable } from "@nestjs/common";
 
-import { createHashedPath } from "../../dam/files/files.utils";
 import { BlobStorageConfig } from "../blob-storage.config";
 import { BLOB_STORAGE_CONFIG } from "../blob-storage.constants";
 import { BlobStorageFileUploadInterface } from "../dto/blob-storage-file-upload.interface";
+import { createHashedPath } from "../utils/create-hashed-path.util";
 import { BlobStorageAzureStorage } from "./azure/blob-storage-azure.storage";
 import { BlobStorageBackendInterface, CreateFileOptions, StorageMetaData } from "./blob-storage-backend.interface";
 import { BlobStorageFileStorage } from "./file/blob-storage-file.storage";
