+++ packages/api/cms-api/src/dam/files/files.utils.ts
@@ -5,7 +5,6 @@ import { unlink } from "fs/promises";
 import got from "got";
 import * as mimedb from "mime-db";
 import os from "os";
-import { sep } from "path";
 import slugify from "slugify";
 import stream from "stream";
 import { promisify } from "util";
@@ -20,8 +19,6 @@ export function slugifyFilename(filename: string, extension: string): string {
     const extensionWithDot = extension.startsWith(".") ? extension : `.${extension}`;
     return `${slugify(filename)}${extensionWithDot}`;
 }
-
-export const createHashedPath = (contentHash: string): string => [contentHash.substr(0, 2), contentHash.substr(2, 2), contentHash].join(sep);
 
 export const calculatePartialRanges = (size: number, range: string): { start: number; end: number; contentLength: number } => {
     let [start, end] = range.replace(/bytes=/, "").split("-") as Array<string | number>;
